allow for demux delay of 150nS and add divider to ila

+++ zynq/copyboot.sh
@@ -5,6 +5,11 @@
 #
 set -e -v
 cd `dirname $0`
+if ! grep 'Xilinx Zynq' /proc/cpuinfo
+then
+    echo "doesn't look like a zynq"
+    exit 1
+fi
 gunzip -c BOOT.BIN.gz | sudo dd of=/boot/BOOTX.BIN bs=4096
 sudo mv /boot/BOOTX.BIN /boot/BOOT.BIN
 sudo reboot
